Move the modal to a component
